requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-telegram-bot==20.7
google-generativeai==0.4.0
duckduckgo-search==4.4.3
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.2
